requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
flask==3.0.0
openai==0.28.1
python-dotenv==1.0.1
httpx>=0.24.1 